# pkg/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoopWorker API",
    "version": "1.0.0",
    "description": "Work-loop engine REST API. Every response uses one envelope: {success, data|error, timestamp, request_id}. Errors carry a stable machine-readable code plus a message that states the cause and the fix. Credentials: send 'X-API-Key: lwk_...' or 'Authorization: Bearer <token>'. A fresh install prints an ephemeral admin API key once at startup; permanent keys come from LOOPWORKER_API_KEYS or POST /api/v1/auth/keys.",
    "contact": { "name": "LoopWorker self-service docs", "url": "http://localhost:19527/api/v1/openapi.json" }
  },
  "servers": [ { "url": "http://localhost:19527", "description": "local default" } ],
  "security": [ { "ApiKeyHeader": [] }, { "BearerAuth": [] } ],
...
  "paths": {
...
    "/api/v1/tasks/{taskID}": {
      "get": {
        "tags": ["tasks"], "operationId": "getTask", "summary": "Read one task",
        "parameters": [ { "$ref": "#/components/parameters/taskID" } ],
        "responses": {
          "200": { "description": "task", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TaskEnvelope" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["tasks"], "operationId": "deleteTask", "summary": "Cancel a task (requires 'write')",
        "parameters": [ { "$ref": "#/components/parameters/taskID" } ],
        "responses": {
          "200": { "description": "cancelled" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
    "/api/v1/workflow/list": {
      "get": { "tags": ["workflow"], "operationId": "listWorkflows", "summary": "Registered workflows",
        "responses": { "200": { "description": "workflows" }, "503": { "$ref": "#/components/responses/Unavailable" } } }
    },
    "/api/v1/workflow/{workflowID}": {
      "get": { "tags": ["workflow"], "operationId": "getWorkflow", "summary": "One workflow with step statuses",
        "parameters": [ { "$ref": "#/components/parameters/workflowID" } ],
        "responses": { "200": { "description": "workflow" }, "404": { "$ref": "#/components/responses/NotFound" } } }
    },
    "/api/v1/workflow/execute": {
      "post": {
        "tags": ["workflow"], "operationId": "executeWorkflow", "summary": "Start a workflow (requires 'execute'); returns 202 and is polled",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["workflow_id"], "additionalProperties": false,
          "properties": { "workflow_id": { "type": "string" } }
        } } } },
        "responses": { "202": { "description": "executing" }, "404": { "$ref": "#/components/responses/NotFound" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/api/v1/workers": {
      "get": { "tags": ["system"], "operationId": "listWorkers", "summary": "Worker pool state",
        "responses": { "200": { "description": "workers" }, "503": { "$ref": "#/components/responses/Unavailable" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": { "type": "apiKey", "in": "header", "name": "X-API-Key", "description": "lwk_ prefixed key; stored as a SHA-256 digest" },
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "lwt", "description": "HMAC-signed token from POST /api/v1/auth/token" }
    },
    "parameters": {
      "taskID": { "name": "taskID", "in": "path", "required": true, "schema": { "type": "string" } },
      "workflowID": { "name": "workflowID", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "schemas": {
      "TaskState": { "type": "string", "enum": ["pending", "queued", "running", "completed", "failed", "cancelled", "retrying", "dead_letter"] },
      "ErrorCode": {
        "type": "string",
        "description": "Stable machine-readable error code. Never changes meaning.",
        "enum": ["INVALID_REQUEST", "UNKNOWN_FIELD", "MALFORMED_JSON", "INPUT_ENCODING_INVALID", "REQUEST_TOO_LARGE",
          "UNAUTHORIZED", "TOKEN_INVALID", "TOKEN_EXPIRED", "FORBIDDEN", "TASK_NOT_FOUND", "TASK_ALREADY_EXISTS",
          "TASK_STATE_CONFLICT", "WORKFLOW_NOT_FOUND", "WORKER_NOT_FOUND", "PLUGIN_NOT_FOUND", "DEPENDENCY_CYCLE",
          "DEPENDENCY_INVALID", "GRAPH_TOO_LARGE", "RATE_LIMITED", "STREAM_LIMIT_REACHED", "QUEUE_FULL", "NOT_FOUND",
          "METHOD_NOT_ALLOWED", "STREAMING_UNSUPPORTED", "SERVICE_UNAVAILABLE", "INTERNAL_ERROR", "TIMEOUT"]
      },
      "ErrorBody": {
        "type": "object",
        "required": ["code", "message", "request_id"],
        "properties": {
          "code": { "$ref": "#/components/schemas/ErrorCode" },
          "message": { "type": "string", "description": "Cause first, then 'Fix: ' with the repair step" },
          "details": { "type": "object", "additionalProperties": true },
          "request_id": { "type": "string" }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": ["success", "error", "timestamp", "request_id"],
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "$ref": "#/components/schemas/ErrorBody" },
          "timestamp": { "type": "string", "format": "date-time" },
          "request_id": { "type": "string" }
        }
      },
      "Task": {
        "type": "object",
        "required": ["id", "type", "state", "priority", "created_at", "retry", "max_retry", "dependencies", "is_agent"],
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "state": { "$ref": "#/components/schemas/TaskState" },
          "priority": { "type": "integer" },
          "priority_name": { "type": "string", "enum": ["low", "normal", "high", "critical", "unknown"] },
          "owner": { "type": "string", "description": "server-assigned subject id of the creating credential" },
          "created_at": { "type": "string", "format": "date-time" },
          "started_at": { "type": "string", "format": "date-time" },
          "ended_at": { "type": "string", "format": "date-time" },
          "input": { "type": "string" },
          "input_encoding": { "type": "string", "enum": ["text", "base64"] },
          "result": { "type": "string" },
          "result_encoding": { "type": "string", "enum": ["text", "base64"] },
          "error": { "type": "string" },
          "retry": { "type": "integer" },
          "max_retry": { "type": "integer" },
          "dependencies": { "type": "array", "items": { "type": "string" } },
          "config": { "type": "object", "additionalProperties": true, "description": "secret-looking values are replaced with ***redacted***" },
          "metadata": { "type": "object", "additionalProperties": { "type": "string" } },
          "is_agent": { "type": "boolean" },
          "agent_config": { "type": "object", "properties": { "system_prompt": { "type": "string" }, "model": { "type": "string" }, "schema": { "type": "string" } } }
        }
      },
      "TaskEnvelope": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/Task" },
          "timestamp": { "type": "string", "format": "date-time" },
          "request_id": { "type": "string" }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": { "description": "401 UNAUTHORIZED / TOKEN_INVALID / TOKEN_EXPIRED", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Forbidden": { "description": "403 FORBIDDEN with the missing permission in details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "NotFound": { "description": "404 TASK_NOT_FOUND / WORKFLOW_NOT_FOUND / NOT_FOUND", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
      "Conflict": { "description": "409 TASK_STATE_CONFLICT / DEPENDENCY_CYCLE", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "Unavailable": { "description": "503 SERVICE_UNAVAILABLE / QUEUE_FULL", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    }
  }
}